Correct project
Add Mnemo08Info script. Correct Mnemo12. Fix Mnemo8 animation
