+++ Assets/Mnemonic schemes/0 Animation/08 Mnemo Animation Weight.anim
@@ -2135,8 +2135,43 @@ AnimationClip:
     floatParameter: 0
     intParameter: 0
     messageOptions: 0
+  - time: 1
+    functionName: SetDestinationPositionIsEntry
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 1
+    functionName: SetVerticalMovementLevelIsBottom
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 4
+    functionName: SetVerticalCoordinateIs5000
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
   - time: 8
     functionName: ContinueCoordinateManipulatorAnimation
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 8
+    functionName: SetVerticalMovementLevelIsTop
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 11
+    functionName: SetVerticalCoordinateIs9500
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
@@ -2172,6 +2207,13 @@ AnimationClip:
     messageOptions: 0
   - time: 38.816666
     functionName: SetGeometryControlTrue
+    data: 
+    objectReferenceParameter: {fileID: 0}
+    floatParameter: 0
+    intParameter: 0
+    messageOptions: 0
+  - time: 38.816666
+    functionName: SetDestinationPositionIsGeometry
     data: 
     objectReferenceParameter: {fileID: 0}
     floatParameter: 0
