+++ Assets/Scripts/infoOnMnemo.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 539633381711a5547b0340706af34969
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
