+++ Assets/Scripts/infoOnMnemo/Mnemo08Info.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6be89dc683b9caa4b9b1ec2384b40ff2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
