+++ Assets/Scripts/infoOnMnemo/Mnemo12Info.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d2a6ee85ffd86f24fa99924c2d8761a5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
